# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Agiletech Test"
  },
  "host": "agiletech-test.herokuapp.com",
  "tags": [
    {
      "name": "Auth",
      "description": "Authen"
    },
    {
      "name": "Post",
      "description": "Post"
    },
    {
      "name": "Galleries",
      "description": "Galleries"
    }
  ],
  "schemes": ["https", "http"],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "description": "",
        "operationId": "authLogin",
        "parameters": [
          {
            "name": "Request body",
            "in": "body",
            "description": "The user name for login",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/AuthLoginBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/auth/refresh-token": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh Token",
        "description": "",
        "operationId": "authRefreshToken",
        "parameters": [
          {
            "name": "Request body",
            "in": "body",
            "description": "",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/AuthRefreshBody"
            }
          }
        ],

        "responses": {
          "200": {
            "description": "successful operation",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/auth/logout": {
      "delete": {
        "tags": ["Auth"],
        "summary": "Logout",
        "description": "",
        "operationId": "authLogout",
        "produces": ["application/json"],
        "parameters": [],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/posts": {
      "get": {
        "tags": ["Post"],
        "summary": "Get Posts",
        "description": "",
        "operationId": "getPosts",
        "produces": ["application/json"],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": true,
            "type": "string",
            "example": "1"
          },
          {
            "name": "title",
            "in": "query",
            "description": "",
            "type": "string",
            "example": "title here"
          },
          {
            "name": "tags",
            "in": "query",
            "description": "",
            "type": "string",
            "example": "Html"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Post"],
        "summary": "Create Posts",
        "description": "",
        "operationId": "createPosts",
        "produces": ["application/json"],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "Request body",
            "in": "body",
            "description": "",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/CreatePostBody"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/posts/{postId}": {
      "patch": {
        "tags": ["Post"],
        "summary": "Edit Post",
        "description": "",
        "operationId": "editPost",
        "produces": ["application/json"],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "type": "string",
            "example": "1"
          },
          {
            "name": "page",
            "in": "body",
            "description": "",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/CreatePostBody"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": ["Post"],
        "summary": "Delete Post",
        "description": "",
        "operationId": "deletePost",
        "produces": ["application/json"],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "type": "string",
            "example": "1"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/tags": {
      "get": {
        "tags": ["Post"],
        "summary": "Get Tags",
        "description": "",
        "operationId": "getTags",
        "produces": ["application/json"],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/galleries": {
      "get": {
        "tags": ["Galleries"],
        "summary": "Get Galleries",
        "description": "",
        "operationId": "getGalleries",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  },
  "definitions": {
    "AuthLoginBody": {
      "type": "object",
      "properties": {
        "username": { "type": "string", "example": "admin" }
      }
    },
    "AuthRefreshBody": {
      "type": "object",
      "properties": {
        "refreshToken": { "type": "string", "example": "refresh token here" }
      }
    },
    "CreatePostBody": {
      "type": "object",
      "properties": {
        "title": { "type": "string", "example": "Lorem" },
        "desciption": { "type": "string", "example": "Lorem" },
        "tags": { "type": "object", "example": ["Html", "Css"] }
      }
    }
  }
}